Refactor GraphQL imports and delete barrel file
- Removed src/gql/index.ts
- Refactored Issue.tsx and CreateIssue.tsx to import mutations directly
- Renamed src/gql/ISSUES_QUERY_.ts to src/gql/ISSUES_FOR_USER_QUERY.ts with a TODO
