--- src/gql/index.ts
@@ -1,11 +1,0 @@
-// TODO: Delete this file
-import { CREATE_ISSUE_MUTATION } from "./CREATE_ISSUE_MUTATION";
-import { UPDATE_ISSUE_STATUS_MUTATION } from "./UPDATE_ISSUE_STATUS_MUTATION";
-import { DELETE_ISSUE_MUTATION } from "./DELETE_ISSUE_MUTATION";
-import { ISSUES_QUERY } from "./ISSUES_QUERY_";
-export {
-  ISSUES_QUERY,
-  CREATE_ISSUE_MUTATION,
-  DELETE_ISSUE_MUTATION,
-  UPDATE_ISSUE_STATUS_MUTATION,
-};
